DB_/requirements.txt
psycopg2-binary
pandas
datasets
scikit-learn